Refactor and rename scheduler agenda injection helper
