+++ packages/shared/src/helpers/index.ts
@@ -1,6 +1,8 @@
 import { withExpect } from '@acala-network/chopsticks-testing'
 import { assert, expect } from 'vitest'
 
+import type { HexString } from '@polkadot/util/types'
+
 const { check, checkEvents, checkHrmp, checkSystemEvents, checkUmp } = withExpect((x: any) => ({
   toMatchSnapshot(msg?: string): void {
     expect(x).toMatchSnapshot(msg)
@@ -54,4 +56,30 @@ export function objectCmp(
       assert(cmp, errorMessage)
     }
   }
+}
+
+/**
+ * Given a PJS client and a hex-encoded extrinsic with a given origin, modify the
+ * `scheduler` pallet's `agenda` storage to execute the extrinsic in the next block.
+ */
+export async function scheduleCallWithOrigin(client: any, call: HexString, origin: any) {
+  const number = (await client.api.rpc.chain.getHeader()).number.toNumber()
+
+  await client.dev.setStorage({
+    Scheduler: {
+      agenda: [
+        [
+          [number + 1],
+          [
+            {
+              call: {
+                Inline: call,
+              },
+              origin: origin,
+            },
+          ],
+        ],
+      ],
+    },
+  })
 }
